Get server working + add dependency injection and change from double to decimal

+++ KnowYourWatts.ClientUI.DTO/Requests/CurrentUsageRequest.cs
@@ -5,5 +5,5 @@ namespace KnowYourWatts.DTO.Requests;
 public sealed class CurrentUsageRequest
 {
     public TariffType TariffType { get; set; }
-    public double CurrentReading { get; set; }
+    public decimal CurrentReading { get; set; }
 }
