+++ KnowYourWatts.ClientUI.DTO/Requests/DailyUsageRequest.cs
@@ -5,5 +5,5 @@ namespace KnowYourWatts.DTO.Requests;
 public class DailyUsageRequest
 {
     public TariffType TariffType { get; set; }
-    public double CurrentReading { get; set; }
+    public decimal CurrentReading { get; set; }
 }
